Remove the Play version from config
Some Play packages have now its own lifecycle. So it isn't guaranteed that all packages have the same Play version. This project is also built against a specific major version of Play, so at this time it doesn't make sense to use a different Play version. The different Play packages are binary compatible against a major version, so it should not be an issue to hardcode the versions of the specific libraries in the build.sbt file.